workers/requirements.txt
# Database (use a version compatible with Python 3.13)
sqlalchemy==2.0.42
# psycopg v3 with prebuilt wheels (works on Windows/Python 3.13 without MSVC build tools)
psycopg[binary]==3.2.3

# Environment & Config
python-dotenv==1.0.0

# Redis queue
redis==7.1.0

# File parsing
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# GAEB parsing (German tender format)
gaeb-xml-parser>=0.2.0

# OCR support
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3

# LLM client
openai>=1.30.0

# HTTP API
fastapi==0.110.0
uvicorn==0.27.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Type checking (optional)
mypy==1.8.0
